refactor: use ReferentName instead of ProjectName
